Augment the WebGPU API in compat
Sometimes it might be useful to augument the WebGPU API
from JavaScript while developing the CTS. A specific usecase
that came up for me was I wanted to write tests that validated
new compat rules but I was unable to easily add that validation
to an actual implementation.

So, instead, I added the validation at a JavaScript level.
This involves tracking the error scope stack and inserting
new validation error results.

This patch is an example of that soluition.

This allowed me to write new compat tests as well as find
out which CTS tests are going to start failing in compat
mode so that I could fix the tests without waiting for
an implementation.

* standalone.ts is used in the browser.
* cmdline.ts is used from dawn.node.