Emulate @interpolate(flat) when hardware uses last as provoking vertex
Ok, I tried to do the emu **in JavaScript** and .... it's effectively impossible.
I think Stephen already recognized this. In JS it's hard because you can't easily
insert some compute pass before a render pass. You instead need to record all of
the render pass commands so that you can then generate a compute pass first, and
only after create the render pass because only one pass encoder can be active at
a time.

It gets impossible because of drawIndirect. You don't know what size index buffer
you'll need since that info is in the indirect buffer and in WebGPU there is no
way to get that info in a performant way. You'd have to asynchronously map the
buffer to read out the the vertex_count to know what size index buffer you need
for rotated indices.

In Dawn, the first part is not a problem. Because you can insert arbitrary work
before the glDrawXXX call. I'm not sure about the 2nd. You'd have to map the
buffer, read out the vertex_count and then allocate the index buffer.

Wrote up notes here:
 https://docs.google.com/document/d/1tORVt_qmayuDl8rDM2QP0-EzeeGxWHzaJ4bEoze7Ncs/edit?usp=sharing&resourcekey=0-hPP1hHUu-gzr2t0smeOj9Q