WGSL textureSample/Bias/Grad tests
Add all the textureSample, textureSampleBias, and textureSampleGrad
tests. cube and cube-array tests with derivatives are currently skipped
or filtered out as the software rasterizer can't handle this case
correctly or at least doesn't match too many GPUs. Rather than increase
the tolerances.

There's a big change to the soft rasterizer to compute derivatives
by setting up 2x2 pixels and computing the derivatives by the
differences between the directions.

Another big change is the software rasterizer, instead of rendering
1 texture, it renders the min and max of all texels sampled.

Yet Another big change is the doTextureCalls function now draws a triangle
to a 1x1 size texture, per call using isntance drawing. This allows us
to control the derivatives on a per call basis. Derviatives are based
on a variable called `derivativeBase`. It just the (fragment position - 0.5)
so (0,0) for our 1x1 pixel. It is then divided by the size of the texture
which means its implicit derivatives will choose mip level 0. This
will allow us in a future PR to multiply it to choose any mip level we
want to test. This means we can probably remove the xxx,derivatives tests
and include derivatives in the main tests. But, as this appears to be
working now, I'd let to check it in.

A future PR will add random or semi-random derviatives to each call.