Try emulating textureGather for 3d textures to workaround GPU bugs
Some GPUs (intel) don't appear to correctly wrap texture coords
around cubemap faces when they should. This code (see diff)
emulates textureGather on top of textureSampleLevel

To do this it converts the cube coord to a 3d normalized coord
where xy are a normalized texture coord and z is a normalized
layer coord

That coord is then multiplied by the size of the texture and
quantized. floor and ceil are called to get the 4 texel coords
that would be "gathered". If they wrap then floor will produce
a negative texel coord (off the current face) and or ceil will
produce a texel coord that equals the size for texture (also
off the face)

We then convert these back to a normalized texture coord which
means the out of range coord will be < 0 or > 1. We convert
that normalized texture coord back to a cube coord. If the
normalized texture coord was < 0 or > 1 then the resulting
cube coord will now point to a different face.

We then need to quantize the result. This is because in an
infinite face plane, the out of bounds cube coord would point
directly to the center of a texel in the plane in started in but
it will be off center in the new plane. Quantizing it gets it
back in the center of the texel on the new plane.

To quantize we convert it back to a 2d normalized coord.
Covert that to texels. Quantize. Conver that back to a normalized
coord. Convert that back to a cube coord.